Removed clock and using sinon for time travel
