--- lib/clock.js
@@ -1,8 +1,0 @@
-'use strict';
-module.exports = () => new Clock();
-
-class Clock {
-  currentMillis() {
-    return new Date().getTime();
-  }
-}
