+++ package.json
@@ -24,6 +24,7 @@
   },
   "devDependencies": {
     "chai": "~3.5.0",
-    "mocha": "~3.1.0"
+    "mocha": "~3.1.0",
+    "sinon": "~1.17.6"
   }
 }
